z now works correctly according to note 3 to chapter 4
i.e. in Myrvold's 12/2017 MS.  This behavior differs from the
algorithm specified in the text on pages 93 and 103, but I think it's
what's intended. I realized that initial zeros don't show up
distinctively, so it wasn't a bug that z always has value 0.1 the
first time x has a  value > 0.5.